paging limit 변수명 통일

+++ src/main/java/com/nts/reservation/service/main/MainProductService.java
@@ -15,14 +15,14 @@ public interface MainProductService {
 	 * @param	start	페이지에 출력할 데이터의 시작 index
 	 * @return	인터페이스의 LIMIT만큼 해당 DTO의 List로 반환
 	 */
-	List<MainProduct> getProducts(long categoryId, long start, int pagingLimit);
+	List<MainProduct> getProducts(int categoryId, int start, int pagingLimit);
 
 	/**
 	 * api/products response에 사용. 카테고리 구분 없음
 	 * @param	start	페이지에 출력할 데이터의 시작 index
 	 * @return	인터페이스의 LIMIT만큼 해당 DTO의 List로 반환
 	 */
-	List<MainProduct> getProducts(long start, int pagingLimit);
+	List<MainProduct> getProducts(int start, int pagingLimit);
 
 	/**
 	 * api/products response에 사용
